Add unit tests for SizeVisitor.

+++ Behavioral/DesignPatterns.Visitor/ISize.cs
@@ -2,7 +2,7 @@ namespace DesignPatterns.Visitor;
 
 public interface ISize
 {
-    double Value { get; }
+    long Value { get; }
     string Unit { get; }
     ISize Accept(IVisitor<ISize, ISize> visitor);
 }
