+++ Behavioral/DesignPatterns.Visitor/Models/Size.cs
@@ -2,7 +2,7 @@ namespace DesignPatterns.Visitor.Models;
 
 public record class Size : ISize
 {
-    public double Value { get; init; }
+    public long Value { get; init; }
     public virtual string Unit { get; init; }
 
     public ISize Accept(IVisitor<ISize, ISize> visitor)
